trying to move hold button creation into climb class so functions can access the hold_buttons buttongroup
